+++ model/acl_group_model.js
@@ -7,20 +7,20 @@
  * @since    0.0.1
  * @version  0.5.0
  */
-var Group = Function.inherits('Alchemy.Model', function AclGroup(options) {
-	AclGroup.super.call(this, options);
-});
+var Group = Function.inherits('Alchemy.Model', 'AclGroup');
 
 /**
  * Constitute the class wide schema
  *
  * @author   Jelle De Loecker   <jelle@develry.be>
  * @since    0.2.0
- * @version  0.2.0
+ * @version  0.7.2
  */
 Group.constitute(function addFields() {
 	this.addField('name', 'String');
+	this.addField('title', 'String');
 	this.addField('weight', 'Number');
+	this.addField('description', 'String');
 });
 
 /**
@@ -28,13 +28,12 @@ Group.constitute(function addFields() {
  *
  * @author   Jelle De Loecker <jelle@develry.be>
  * @since    0.3.0
- * @version  0.4.0
+ * @version  0.7.2
  */
 Group.constitute(function chimeraConfig() {
 
 	var list,
-	    edit,
-	    view;
+	    edit;
 
 	if (!this.chimera) {
 		return;
@@ -43,18 +42,16 @@ Group.constitute(function chimeraConfig() {
 	// Get the list group
 	list = this.chimera.getActionFields('list');
 
+	list.addField('title');
 	list.addField('name');
 	list.addField('weight');
+	list.addField('description');
 
 	// Get the edit group
 	edit = this.chimera.getActionFields('edit');
-
+	
+	edit.addField('title');
 	edit.addField('name');
 	edit.addField('weight');
-
-	// Get the view group
-	view = this.chimera.getActionFields('view');
-
-	view.addField('name');
-	view.addField('weight');
+	edit.addField('description');
 });