+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.2 (WIP)
 
 * Move User's beforeSave logic to a method instead of an event
+* Add title & description field to ACL Group model
 
 ## 0.7.1 (2021-09-12)